backend/requirements.txt
argcomplete==3.6.3
asgiref==3.11.0
bcrypt==5.0.0
boto3==1.42.19
botocore==1.42.19
certifi==2025.11.12
cfn-flip==1.3.0
charset-normalizer==3.4.4
click==8.3.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
durationpy==0.10
hjson==3.1.0
idna==3.11
jmespath==1.0.1
MarkupSafe==3.0.3
placebo==0.9.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.6
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
zappa==0.61.3
